MGMT-15963: Upgrade dockefiles to use centos:stream9 and build with podman
The current Dockerfile builds in this project are broken due to recent deprecation centos images.

This PR aims to address that by upgrading the images we use.
Additionally the build system is made more consistent with the podman compliant build that we use for assisted.
The host podman is passed to skipper so that the container may use podman-remote to build the project.